third_party/WebKit/LayoutTests/http/tests/security/isolatedWorld/events-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://127.0.0.1:8000/security/isolatedWorld/anyfile. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access. The response had HTTP status code 404.


Text attribute event listeners should always fire in main world: PASS
Function attribute event listeners should fire in the world where they were created: PASS
XHR attribute event listeners should fire in the world where they were created: PASS
